requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
python-multipart==0.0.9
pandas==2.2.2
openpyxl==3.1.4
xlrd==2.0.1
psycopg2-binary==2.9.9
